finish move to activity, delete old files
