+++ src/main/java/com/applicaster/Constants.java
@@ -1,0 +1,7 @@
+package com.applicaster;
+
+public class Constants {
+
+    public static int ANIMATION_DURATION = 400;
+
+}
